chore(storybook): Restructure the public page template layout modules
Rename config to commonMeta, Layout to PageLayout, and FormLayout to FormPageLayout, and inline the App Header into the Page Layout, which was its only consumer.

Let the Page Layout render the Page component, like the Form Page Layout already did, so each layout owns the whole page structure and the decorators using them shrink to a single element.

Extract a formPageLayoutDecorator in the Form Flow stories, which repeated the same decorator four times.
